+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 -----
 
 - Add a pull to refresh to the podcast list and profile view (#239)
+- Add a helpful alert if the user is signed out in the background (#251)
 
 7.22
 -----